+++ Samples/Compositor/include/Compositor.h
@@ -359,7 +359,7 @@ void Sample_Compositor::itemSelected(OgreBites::SelectMenu* menu)
 	}
 
 	mTrayMgr->getWidget("DebugRTTPanel")->show();
-	mTrayMgr->moveWidgetToTray("DebugRTTPanel", TL_TOPRIGHT, mTrayMgr->getNumWidgets(TL_TOPRIGHT) - 1);
+	mTrayMgr->moveWidgetToTray("DebugRTTPanel", TL_TOPRIGHT, (unsigned long)mTrayMgr->getNumWidgets(TL_TOPRIGHT) - 1);
 	StringVector parts = StringUtil::split(menu->getSelectedItem(), ";");
 	mDebugTextureTUS->setContentType(TextureUnitState::CONTENT_COMPOSITOR);
